+++ src/plot/layer/geom/arrow.rs
@@ -43,10 +43,6 @@ impl GeomTrait for Arrow {
     fn supports_aggregate(&self) -> bool {
         true
     }
-
-    fn aggregate_slots(&self) -> &'static [u8] {
-        &[1, 2]
-    }
 }
 
 impl std::fmt::Display for Arrow {