+++ src/plot/layer/geom/point.rs
@@ -42,10 +42,6 @@ impl GeomTrait for Point {
     fn supports_aggregate(&self) -> bool {
         true
     }
-
-    fn aggregate_slots(&self) -> &'static [u8] {
-        &[1, 2]
-    }
 }
 
 impl std::fmt::Display for Point {